Move QLPreviewPanel logic to ConversionCompletedView
+ add zoom image
